requirements.txt
discord.py
python-dotenv
psycopg2-binary
requests
Flask
SQLAlchemy
gunicorn
psutil
